More refactors, test review changes
