+++ waspc/data/packages/wasp-config/__tests__/spec/testFixtures.ts
@@ -30,10 +30,7 @@ export function getApp(scope: ConfigType): Config<TsAppSpec.App> {
         wasp: { version: "^0.16.3" },
         title: "Mock App",
         head: ['<link rel="icon" href="/favicon.ico" />'],
-        // TODO: should we fill the parts here? The old approach didn't have this problem
-        // Because this test only cared about the initial object passed to `new App`. The
-        // whole system was tested in integration tests.
-        parts: [],
+        parts: [getPage("full"), getQuery("full")],
       });
     default:
       assertUnreachable(scope);
